added some comments

+++ client/src/main/java/client/group12/vegetarianMeal/mealAPI.java
@@ -24,13 +24,13 @@ public class mealAPI {
         ObjectMapper objectMapper = new ObjectMapper();
         JsonNode rootNode = objectMapper.readTree(new URL("https://opendata.socrata.com/resource/8nz9-yn2p.json"));
 
-
+        //iterates over all the elements in the JsonNode and stores the data we need (food name, co2 emmision per serving and the size of the serving)
         Iterator<JsonNode> elements = rootNode.elements();
-        while(elements.hasNext()){
+        while(elements.hasNext()) {
             JsonNode n = elements.next();
             String food = n.get("food").asText();
             int co2Node = n.get("grams_co2e_per_serving").asInt();
-            int sizeNode = n.get("serving_size").asInt();
+            double sizeNode = n.get("serving_size").asInt();
 
             mealList.add(new Meal(food, co2Node, sizeNode));
             elements.next();
@@ -41,5 +41,6 @@ public class mealAPI {
             System.out.println(mealList.get(i).getCo2());
             System.out.println(mealList.get(i).getServingSize() + "\n");
         }
+
     }
 }
